Switch to using LigtBulbSystem
