--- Content.Client/Light/Visualizers/LightBulbVisualizerComponent.cs
@@ -1,8 +1,0 @@
-using Content.Shared.Light;
-using Robust.Client.GameObjects;
-
-namespace Content.Client.Light.Visualizers;
-
-[RegisterComponent]
-public sealed class LightBulbVisualizerComponent : Component
-{}
