+++ Content.Server/Light/Components/PoweredLightComponent.cs
@@ -1,7 +1,7 @@
 using System.Threading;
 using Content.Server.Light.EntitySystems;
 using Content.Shared.Damage;
-using Content.Shared.Light;
+using Content.Shared.Light.Component;
 using Content.Shared.MachineLinking;
 using Robust.Shared.Audio;
 using Robust.Shared.Containers;
